Move timeline service to client-lib
